refactor(pages): render Convex-backed /live + /admin client-only (ssr:false)
Both pages are entirely Convex/auth-driven, so their interactive bodies now
load via next/dynamic with ssr:false, leaving only a lightweight shell to
prerender. This makes them structurally immune to prerender-time crashes on a
missing ConvexProvider (the class of failure that broke /talks in CI) and drops
the "Connecting…" flash from the static HTML.

- components/live/LiveRoom.tsx: extracted the live room (+ not-configured
  fallback) out of pages/live.
- components/admin/AdminDashboard.tsx: extracted the cockpit (AdminGate +
  dashboard) out of pages/admin.

No behaviour change; useQuery is client-only regardless, so nothing was ever
server-rendered here. Verified: typecheck, lint, and a production build with
NEXT_PUBLIC_CONVEX_URL unset (the CI condition) all green — /live and /admin
prerender as clean static shells.
